Move ContainerProperties to a Top Level Class
