+++ spring-kafka/src/main/java/org/springframework/kafka/listener/config/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Container configuration.
+ */
+package org.springframework.kafka.listener.config;
